Changed location of services object
With this latest commmit the `services` object has been moved into its
own JS file to later expansion.
Renamed ODH.js to Services.js

Misc Changes:
- Combined mantine imports for cleanliness
- Removed right and left padding on the services container to match that
  of the clusters
